Update create.tsx
Allow float numbers as price to be minted

In this updated version of the function, the price value is multiplied by 1,000,000 to convert it to micro-units before being passed to the mintWithMetadata method. When you retrieve the value later, you can convert it back to a floating-point number by dividing it by 1,000,000.